+++ apps/blog/README.md
@@ -1,19 +1,19 @@
-# Blog5280
+# Blog
 
 **TODO: Add description**
 
 ## Installation
 
 If [available in Hex](https://hex.pm/docs/publish), the package can be installed
-by adding `blog5280` to your list of dependencies in `mix.exs`:
+by adding `blog` to your list of dependencies in `mix.exs`:
 
 ```elixir
 def deps do
-  [{:blog5280, "~> 0.1.0"}]
+  [{:blog, "~> 0.1.0"}]
 end
 ```
 
 Documentation can be generated with [ExDoc](https://github.com/elixir-lang/ex_doc)
 and published on [HexDocs](https://hexdocs.pm). Once published, the docs can
-be found at [https://hexdocs.pm/blog5280](https://hexdocs.pm/blog5280).
+be found at [https://hexdocs.pm/blog](https://hexdocs.pm/blog).
