+++ apps/blog/config/config.exs
@@ -10,11 +10,11 @@ use Mix.Config
 
 # You can configure for your application as:
 #
-#     config :blog5280, key: :value
+#     config :blog, key: :value
 #
 # And access this configuration in your application as:
 #
-#     Application.get_env(:blog5280, :key)
+#     Application.get_env(:blog, :key)
 #
 # Or configure a 3rd-party app:
 #
