+++ apps/blog/lib/blog/application.ex
@@ -1,4 +1,4 @@
-defmodule Blog5280.Application do
+defmodule Blog.Application do
   # See http://elixir-lang.org/docs/stable/elixir/Application.html
   # for more information on OTP Applications
   @moduledoc false
@@ -10,13 +10,13 @@ defmodule Blog5280.Application do
 
     # Define workers and child supervisors to be supervised
     children = [
-      # Starts a worker by calling: Blog5280.Worker.start_link(arg1, arg2, arg3)
-      worker(Blog5280, []),
+      # Starts a worker by calling: Blog.Worker.start_link(arg1, arg2, arg3)
+      worker(Blog, []),
     ]
 
     # See http://elixir-lang.org/docs/stable/elixir/Supervisor.html
     # for other strategies and supported options
-    opts = [strategy: :one_for_one, name: Blog5280.Supervisor]
+    opts = [strategy: :one_for_one, name: Blog.Supervisor]
     Supervisor.start_link(children, opts)
   end
 end
