+++ apps/blog/priv/blog/2015-01-09-fitbit-device-integration.md
