+++ apps/blog/priv/blog/2015-02-15-entrepreneurship-101.md
