+++ apps/web5280/web/controllers/page_controller.ex
@@ -2,7 +2,7 @@ defmodule Web5280.PageController do
   use Web5280.Web, :controller
 
   def index(conn, _params) do
-    {:ok, blog_posts} = Blog5280.list()
+    {:ok, blog_posts} = Blog.list()
     render conn, "index.html", blog_posts: blog_posts
   end
 end
